+++ src/containers/catalog-filter/index.js
@@ -1,0 +1,45 @@
+import {memo, useCallback, useMemo} from "react";
+import useStore from "../../store/use-store";
+import useSelector from "../../store/use-selector";
+import Select from "../../components/select";
+import Input from "../../components/input";
+import SideLayout from "../../components/side-layout";
+
+function CatalogFilter() {
+
+  const store = useStore();
+
+  const select = useSelector(state => ({
+    sort: state.catalog.params.sort,
+    query: state.catalog.params.query,
+  }));
+
+  const callbacks = {
+    // Сортировка
+    onSort: useCallback(sort => store.actions.catalog.setParams({sort}), [store]),
+    // Поиск
+    onSearch: useCallback(query => store.actions.catalog.setParams({query, page: 1}), [store]),
+    // Сброс
+    onReset: useCallback(() => store.actions.catalog.resetParams(), [store]),
+  };
+
+  const options = {
+    sort: useMemo(() => ([
+      {value: 'order', title: 'По порядку'},
+      {value: 'title.ru', title: 'По именованию'},
+      {value: '-price', title: 'Сначала дорогие'},
+      {value: 'edition', title: 'Древние'},
+    ]), [])
+  };
+
+  return (
+    <SideLayout padding='medium'>
+      <Select options={options.sort} value={select.sort} onChange={callbacks.onSort}/>
+      <Input value={select.query} onChange={callbacks.onSearch} placeholder={'Поиск'}
+             delay={1000}/>
+      <button onClick={callbacks.onReset}>Сброс</button>
+    </SideLayout>
+  )
+}
+
+export default memo(CatalogFilter);
