+++ src/containers/catalog-list/index.js
@@ -1,0 +1,46 @@
+import {memo, useCallback} from "react";
+import useStore from "../../store/use-store";
+import useSelector from "../../store/use-selector";
+import Item from "../../components/item";
+import List from "../../components/list";
+import Pagination from "../../components/pagination";
+import Spinner from "../../components/spinner";
+
+function CatalogList() {
+  const store = useStore();
+
+  const select = useSelector(state => ({
+    list: state.catalog.list,
+    page: state.catalog.params.page,
+    limit: state.catalog.params.limit,
+    count: state.catalog.count,
+    waiting: state.catalog.waiting,
+  }));
+
+  const callbacks = {
+    // Добавление в корзину
+    addToBasket: useCallback(_id => store.actions.basket.addToBasket(_id), [store]),
+    // Пагинация
+    onPaginate: useCallback(page => store.actions.catalog.setParams({page}), [store]),
+    // генератор ссылки для пагинатора
+    makePaginatorLink: useCallback((page) => {
+      return `?${new URLSearchParams({page, limit: select.limit, sort: select.sort, query: select.query})}`;
+    }, [select.limit, select.sort, select.query])
+  }
+
+  const renders = {
+    item: useCallback(item => (
+      <Item item={item} onAdd={callbacks.addToBasket} link={`/articles/${item._id}`}/>
+    ), [callbacks.addToBasket]),
+  };
+
+  return (
+    <Spinner active={select.waiting}>
+      <List list={select.list} renderItem={renders.item}/>
+      <Pagination count={select.count} page={select.page} limit={select.limit}
+                  onChange={callbacks.onPaginate} makeLink={callbacks.makePaginatorLink}/>
+    </Spinner>
+  );
+}
+
+export default memo(CatalogList);
