+++ src/store/exports.js
@@ -2,3 +2,4 @@ export { default as basket } from './basket';
 export { default as catalog } from './catalog';
 export { default as modals } from './modals';
 export { default as article } from './article';
+export { default as locale } from './locale';
