+++ src/store/locale/index.js
@@ -1,0 +1,19 @@
+import StoreModule from "../module";
+
+class LocaleState extends StoreModule {
+
+  /**
+   * Начальное состояние
+   */
+  initState() {
+    return {
+      lang: 'ru'
+    };
+  }
+
+  setLang(lang) {
+    this.setState({lang}, 'Установлена локаль');
+  }
+}
+
+export default LocaleState;
